% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{add_colors}
\alias{add_colors}
\title{Add color scheme to a \code{ggm} object}
\usage{
add_colors(object, palette, mode = "light")
}
\arguments{
\item{palette}{Color palette options for leads as below:
\itemize{
\item \strong{material} - material color theme
\item \strong{bw} - black or white color theme based on light/dark mode
}}

\item{mode}{Adjust color settings for either light or dark mode}
}
\value{
Returns an updated \code{ggm} object
}
\description{
Add color scheme to a \code{ggm} object
}
