% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{add_intervals}
\alias{add_intervals}
\title{Add intervals}
\usage{
add_intervals(object, intervals = TRUE, channel, minimum_interval = 100)
}
\arguments{
\item{intervals}{The choice of whether interval data will be included. An
annotation channel must be present, otherwise nothing will be plotted. This
argument allows several choices.
\itemize{
\item \strong{TRUE}: all intervals will be annotated (default option)
\item \strong{integer}: an integer vector that represents the indexed intervals that
should be annotated. If NULL, no intervals will be annotated. For example,
if their are 5 beats, than there are 4 intervals between them that can be
labeled. They can be referenced by their index, e.g. \code{intervals = c(2)} to
reference the 2nd interval in a 5 beat range.
}}
}
\description{
Add intervals
}
