% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{age_group_custom}
\alias{age_group_custom}
\title{Create age groups}
\usage{
age_group_custom(x, by = 5, max_age = 60, first_5_different = T,
  as_factor = F, levels_custom = NA, sepa = "-")
}
\description{
\code{age_group_custom} creates age groups with given parameters (called by multiple functions)
}
