% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{annotations}
\alias{annotations}
\alias{draw_boundary_mask}
\title{Add annotations}
\usage{
draw_boundary_mask(object)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Annotations are labels for specific points or samples within a signal. They
can be semantic, in that they may represent the boundary of a region of the
signal, or just an individual peak. They are stored as a WFDB-compatible
annotation file built into a \code{ggm} object.
}
