% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{c.windowed}
\alias{c.windowed}
\title{Concatenate windowed objects}
\usage{
\method{c}{windowed}(...)
}
\arguments{
\item{...}{windowed objects to concatenate}
}
\value{
A windowed object containing all the elements of the input objects
}
\description{
Concatenate windowed objects
}
