% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows-class.R
\name{c.windows}
\alias{c.windows}
\title{Concatenate \code{windows} objects}
\usage{
\method{c}{windows}(...)
}
\arguments{
\item{...}{\code{windows} objects to concatenate.}
}
\value{
A \code{windows} object containing all the elements of the inputs.
}
\description{
Concatenate \code{windows} objects
}
