% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwave.R
\name{calculate_apen}
\alias{calculate_apen}
\title{Calculate Approximate Entropy (ApEn) of a time series}
\usage{
calculate_apen(x, m = 3, r = NULL)
}
\arguments{
\item{x}{Numeric vector of the time series}

\item{m}{Embedding dimension (sample size), default is 3}

\item{r}{Tolerance (threshold), default is 3.5 * sd(x)}
}
\value{
Approximate Entropy value
}
\description{
Calculate Approximate Entropy (ApEn) of a time series
}
\references{
Pincus, S. M. (1991). Approximate entropy as a measure of system
complexity. Proceedings of the National Academy of Sciences, 88(6),
2297-2301.
}
