% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{census}
\alias{census}
\title{Census of genealogical data}
\usage{
census(df, from, to = NA, subset_df = T, born_in_RN = F,
  born_bv = F, gen_age = T, women_only = F, men_only = F, by = 5,
  update_relatives = T, show_messages = F, subset_on_DoB = F,
  min_DoB, max_DoB)
}
\description{
\code{census} generates census of genealogical data (includes more variables than pseudo_census function)
}
