% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.R
\name{cm}
\alias{cm}
\title{Circular Regression}
\usage{
cm(
  formula,
  type,
  data,
  units,
  tidy = FALSE,
  conf.int = TRUE,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{formula}{Standard formula with single dependent variable and multiple
independent variables}

\item{type}{describes the right-hand side of the formula as being either "linear"
or "circular" distributions}

\item{data}{Data set that contains the named terms from the formula}

\item{units}{The type of units that the circular data should be calculated in
(e.g. radians, degrees)}

\item{tidy}{If the results should be in a text block or should be in a tidy
table. If FALSE, then the confidence arguments are not needed/used.}
}
\description{
To be used similar to the \code{stats::lm()} function. This function assumes that
the dependent variable has a circular pattern (e.g. time, angles, directions).
The independent variable may either all be circular, or they may all be linear
(but not both).
}
\details{
Other standards for these types of models have been preset to simplify
modeling. This function requires installation of the \code{{circular}} package for
the back-end of this function.
}
\examples{
formula = el_svg ~ lab_hba1c + bmi
type = "linear"
data = card::geh
units = "degrees"
raw <- cm(formula = formula, type = type, data = data, units = units)
tidy <- cm(formula = formula, type = type, data = data, units = units, tidy = TRUE)

}
