% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{color_channels}
\alias{color_channels}
\title{Modify channel colors}
\usage{
color_channels(x, palette, mode = "light")
}
\arguments{
\item{x}{\if{html}{\out{<character>}} Vector of names of requsted ECG or EGM leads}

\item{palette}{\if{html}{\out{<character>}} Color palette options for leads as below:
\itemize{
\item \strong{material} - material color theme
\item \strong{bw} - black or white color theme based on light/dark mode
}}

\item{mode}{\if{html}{\out{<character>}} Adjust color settings for either light or dark mode}
}
\value{
Character vector of hex code colors based on the selected palette and
light/dark mode
}
\description{
Modify channel colors
}
