% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{colors}
\alias{colors}
\alias{theme_egm}
\alias{theme_egm_light}
\alias{theme_egm_dark}
\title{Theming and color options for \code{ggm} objects}
\usage{
theme_egm(background = c("default", "dark", "light"))

theme_egm_light()

theme_egm_dark()
}
\arguments{
\item{background}{A background style to apply. Use \code{"default"} for the
standard appearance, \code{"dark"} for a black canvas suited to light traces,
or \code{"light"} for a white canvas suited to dark traces.}
}
\value{
A \code{ggm} object, with inheritance similar to
\code{\link[ggplot2:theme_minimal]{ggplot2::theme_minimal()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The general purpose is to improve visualization of electrical signals. There
is a pattern of colors that are generally given from different recording
software, and they can be replicated to help improve visibility.
}
