% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{describe_wfdb2}
\alias{describe_wfdb2}
\title{Describes signals based on WFDB-formatted files}
\usage{
describe_wfdb2(record, location = ".")
}
\arguments{
\item{record}{String that will be used to name the WFDB record. Cannot
include extensions, and is not a filepath. alphanumeric characters are
acceptable, as well as hyphens (-) and underscores (_)}

\item{location}{The directory that the target record is located within. As
this is related to the \href{https://physionet.org}{PhysioNet}, using the
location name \code{mitdb} will access the online directory for the MIT
Database.}
}
\description{
Describes signals based on WFDB-formatted files
}
