% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-detectors.R
\name{detect_surface_beats}
\alias{detect_surface_beats}
\title{ECG waveform detection}
\usage{
detect_surface_beats(
  record,
  record_dir = ".",
  detector,
  wfdb_path = getOption("wfdb_path"),
  ...
)
}
\arguments{
\item{record}{String that will be used to name the WFDB record. Cannot
include extensions, and is not a filepath. alphanumeric characters are
acceptable, as well as hyphens (-) and underscores (_)}

\item{record_dir}{File path of directory that should be used read and write
files. Defaults to current directory.}

\item{wfdb_path}{Path that leads to installed \code{wfdb} software package on
device. Needs to be directly set using \code{set_wfdb_path()}. Obtained from the
system options on loading of the package, \code{getOption('wfdb_path')}}
}
\value{
Creates a WFDB-compatible annotation file
}
\description{
ECG waveform detection
}
\examples{
record <- "300"
detector <- "ecgpuwave"
}
