% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{do_interpolation}
\alias{do_interpolation}
\title{Helper function to apply interpolation}
\usage{
do_interpolation(original_samples, original_values, new_samples, method)
}
\description{
Helper function to apply interpolation
}
\keyword{internal}
