% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps.R
\name{eps_retrieval}
\alias{eps_retrieval}
\alias{get_frequency}
\alias{get_voltage}
\alias{get_source}
\alias{get_label}
\alias{get_color}
\title{Get components of EPS data}
\usage{
get_frequency(x)

get_voltage(x)

get_source(x)

get_label(x)

get_color(x)
}
\arguments{
\item{x}{signal data from \code{eps} class}
}
\value{
Request value based on attribute that is being called
}
\description{
Get components of EPS data
}
