% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final}
\alias{final}
\title{EGM-generated genealogical dataset of Rio Negro (deduplicated).}
\format{A data frame with 3566 rows and 11 variables: \describe{
  \item{ind_id}{unique universal id}
  \item{DoB}{date of birth}
  \item{DoD}{date of death}
  \item{sex}{sex, numeric}
  \item{sex2}{sex, string}
  \item{alive}{status in 2015}
  \item{father}{father id}
  \item{mother}{mother id}
  \item{children}{children id, comma separated}
  \item{spouses}{spouses id, comma separated}
  \item{id_cuest}{id of original EGM Questionnaire}
  }}
\source{
Data collected by author in Guatemala (2015-2016).
}
\usage{
final
}
\description{
The dataset contains the deduplicated records and relational fields including
all members of the population of Rio Negro.
}
\keyword{datasets}
