% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{find_major_peaks}
\alias{find_major_peaks}
\title{Finds peaks that represent near-field signal}
\usage{
find_major_peaks(x, hz = 1000L, sensitivity = 0.2)
}
\arguments{
\item{hz}{the frequency that the signal was sampled on. Default is 1000/sec.}
}
\description{
Finds peaks that represent near-field signal
}
