% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{format.windowed}
\alias{format.windowed}
\title{Format a windowed object for printing}
\usage{
\method{format}{windowed}(x, ...)
}
\arguments{
\item{x}{A windowed object}

\item{...}{Additional arguments passed to methods}
}
\value{
Invisibly returns x
}
\description{
Format a windowed object for printing
}
