% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows-class.R
\name{format.windows}
\alias{format.windows}
\title{Format a \code{windows} object for printing}
\usage{
\method{format}{windows}(x, ...)
}
\arguments{
\item{x}{A \code{windows} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Invisibly returns x.
}
\description{
Format a \code{windows} object for printing
}
