% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pedigree}
\alias{get_pedigree}
\title{Get pedigree object from EGM data}
\usage{
get_pedigree(famid, ind.q, final, plot = T, highlight = NA)
}
\description{
\code{get_pedigree} shows pedigree object that can be plotted.
}
