% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egm-plots.R
\name{gg_egm}
\alias{gg_egm}
\title{Visualilzation of EGM using \code{ggplot}}
\usage{
gg_egm(data, frequency, channels, time_frame = NULL, ...)
}
\arguments{
\item{data}{Long-format data for visualization. It is a combined format that includes both signal data and header information. Can be generated by the \code{\link[=melt_egm]{melt_egm()}} function.}

\item{frequency}{Integer value of the recording frequency (in Hertz).}

\item{channels}{Character vector of which channels to use. Can give either
the channel label (e.g "CS 1-2") or the recording device/catheter type (e.g
"His" or "ECG").}

\item{time_frame}{A time range that should be displaced given in the format
of a vector with a length of 2. The left value is the start, and right
value is the end time. This is given in seconds (decimals may be used).}
}
\description{
Visualilzation of EGM using \code{ggplot}
}
