% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{ggm}
\alias{ggm}
\title{Visualization of EGMs using \code{ggplot}}
\usage{
ggm(data, channels, time_frame = NULL, ...)
}
\arguments{
\item{data}{Data of the \code{egm} class, which inclues header (meta) and signal
information together.}

\item{channels}{Character vector of which channels to use. Can give either
the channel label (e.g "CS 1-2") or the recording device/catheter type (e.g
"His" or "ECG").}

\item{time_frame}{A time range that should be displaced given in the format
of a vector with a length of 2. The left value is the start, and right
value is the end time. This is given in seconds (decimals may be used).}

\item{annotation_channel}{Identifies which of the selected channels will be
used for annotation. This channel will be used for the analysis of
intervals, peaks, etc. If NULL, no annotations will be made.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
