% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ind.q}
\alias{ind.q}
\title{EGM-generated genealogical dataset of Rio Negro (with duplicates).}
\format{A data frame with 5803 rows and 3 variables: \describe{
  \item{full_id}{unique individual id including id of EGM questionnaire and
  line number in 'Individuals Module'}
  \item{idall}{unique universal id in 'final' dataset}
  \item{h_id}{id of original EGM Questionnaire}
  }}
\source{
Data collected by author in Guatemala (2015-2016).
}
\usage{
ind.q
}
\description{
The dataset contains all the records recorded in EGM interviews in Rio Negro,
including duplicates. Records are grouped by the EGM questionnaires in which
they were originally recorded.
}
\keyword{datasets}
