% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{install_wfdb}
\alias{install_wfdb}
\title{Simple method to load \code{wfdb} and its dependencies}
\usage{
install_wfdb(method = "auto", conda = "auto")
}
\description{
Simple method to load \code{wfdb} and its dependencies
}
