% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals}
\alias{intervals}
\alias{find_major_peaks}
\alias{find_minor_peaks}
\title{Finds peaks that represent near-field signal}
\usage{
find_major_peaks(x, hz = 1000L, sensitivity = 0.2)

find_minor_peaks(x, hz = 1000L, sensitivity = 0.2, specificity = 0.8)
}
\arguments{
\item{hz}{the frequency that the signal was sampled on. Default is 1000/sec.}
}
\description{
Finds peaks that represent near-field signal

Find minor peaks that may be present as near-field signal
}
