% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{is_template}
\alias{is_template}
\title{Test whether an object is a template}
\usage{
is_template(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A single logical value.
}
\description{
Test whether an object is a template
}
