% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows-extract.R
\name{is_window_strategy}
\alias{is_window_strategy}
\title{Test whether an object is a window strategy}
\usage{
is_window_strategy(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A single logical value.
}
\description{
Test whether an object is a window strategy
}
