% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{is_windowed}
\alias{is_windowed}
\title{Test if an object is a windowed object}
\usage{
is_windowed(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE if x is a windowed object, FALSE otherwise
}
\description{
Test if an object is a windowed object
}
