% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{lapply.windowed}
\alias{lapply.windowed}
\title{Apply a function to each element of a windowed object}
\usage{
lapply.windowed(X, FUN, ...)
}
\arguments{
\item{X}{A windowed object}

\item{FUN}{A function to apply to each element}

\item{...}{Additional arguments passed to FUN}
}
\value{
A list of the results of applying FUN to each element of X,
or a new windowed object if all results are egm objects
}
\description{
Apply a function to each element of a windowed object
}
