% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorcardiogram.R
\name{magnitudes}
\alias{magnitudes}
\title{Spatial magnitude of each vector in a loop}
\usage{
magnitudes(xyz)
}
\arguments{
\item{xyz}{A numeric matrix with columns \code{X}, \code{Y}, and \code{Z}.}
}
\value{
A numeric vector of magnitudes.
}
\description{
Spatial magnitude of each vector in a loop
}
\keyword{internal}
