% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{melt_egm}
\alias{melt_egm}
\title{Long Format of Header and Signal Data}
\usage{
melt_egm(header, signal, format, ...)
}
\arguments{
\item{header}{A header table that contains key information about the overall
signal in corresponding channels. Its assumed that the frequency of signal
sampling is equivalent across channels.}

\item{signal}{A \code{data.frame} (or similar) that contains the necessary signal.
Each channel will be labeled using the header information (if unnamed).}

\item{format}{The source of EGM information for specific formats. Support
formats include...
\itemize{
\item \strong{lspro}: The LabSystem PRO from Boston Scientific \emph{ascii.txt} format
}}
}
\description{
Long Format of Header and Signal Data
}
