% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paradata}
\alias{paradata}
\title{Contextual information on EGM interviews.}
\format{A data frame with 109 rows and 3 variables: \describe{
  \item{h_id}{id of original EGM Questionnaire}
  \item{respondent1}{line number of main respondent in
  'Individuals Module' of EGM questionnaire}
  \item{no_u_members}{total number of records in EGM questionnaire}
  }}
\source{
Data collected by author in Guatemala (2015-2016).
}
\usage{
paradata
}
\description{
The dataset contains information on the EGM interviews carried out in Rio
Negro.
}
\keyword{datasets}
