% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{print.windowed}
\alias{print.windowed}
\title{Print a windowed object}
\usage{
\method{print}{windowed}(x, ...)
}
\arguments{
\item{x}{A windowed object}

\item{...}{Additional arguments passed to methods}
}
\value{
Invisibly returns x
}
\description{
Print a windowed object
}
