% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows-class.R
\name{print.windows}
\alias{print.windows}
\title{Print a \code{windows} object}
\usage{
\method{print}{windows}(x, ...)
}
\arguments{
\item{x}{A \code{windows} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Invisibly returns x.
}
\description{
Print a \code{windows} object
}
