% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pseudo_census}
\alias{pseudo_census}
\title{Conduct 'pseudo-census' of EGM data}
\usage{
pseudo_census(year, df, cuts, labels)
}
\description{
\code{pseudo_census} extracts and tallies cross-sectional population data by age and sex.
}
