% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{read_annotation2}
\alias{read_annotation2}
\title{Read in WFDB-compatible annotation files}
\usage{
read_annotation2(record, annotator, location, ...)
}
\arguments{
\item{annotator}{A character string (usually 3 letters) that represents the
type of annotation file or algorithm that generated it. These are the
naming conventions used in the WFDB toolkits}
}
\value{
Annotation file
}
\description{
Read in WFDB-compatible annotation files
}
\section{Annotation files}{
The types of annotations that are supported are described below:
\itemize{
\item atr = manually reviewed and corrected reference annotation files
\item ann = general annotator file
}
}

