% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb.R
\name{read_wfdb}
\alias{read_wfdb}
\title{Read in WFDB-compatible signal files}
\usage{
read_wfdb(record, location = ".", channels = NULL, ...)
}
\arguments{
\item{record}{String that will be used to name the WFDB record. Cannot
include extensions, and is not a filepath. alphanumeric characters are
acceptable, as well as hyphens (-) and underscores (_)}

\item{location}{The directory that the target record is located within. As
this is related to the \href{https://physionet.org}{PhysioNet}, using the
location name \code{mitdb} will access the online directory for the MIT
Database.}
}
\value{
Returns a list of two objects. The first object is a matrix of
signals, with each column representing a specific channel. Each row is a
sample point. The second object is a summary of the header field, with
information on channel names/number, signal length, sampling frequency,
etc.
}
\description{
\code{read_wfdb()} reads in signal files based on the \code{rdsamp} function provided
by the original WFDB toolbox. It requires a signal file, a header file, and
an annotation file (if applicable).
}
