% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythm-package.R
\docType{package}
\name{rhythm-package}
\alias{rhythm}
\alias{rhythm-package}
\title{rhythm: Management of electrocardiogram and electrogram data}
\description{
Importing rhythms and their analysis in R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/rhythm}
  \item Report bugs at \url{https://github.com/asshah4/rhythm/issues}
}

}
\author{
\strong{Maintainer}: Anish Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
