% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation}
\alias{segmentation}
\alias{segment}
\alias{segment_sinus_beats}
\title{Segmentation of electrical signal by wave specifications}
\usage{
segment(object, by, pad)

segment_sinus_beats(object)
}
\arguments{
\item{object}{Object of the \code{egm} class, which includes header, signal
information, and annotation information.}

\item{by}{Character vector naming waveform type to segment by. Options
include the following types:
\itemize{
\item sinus = Will call \code{\link[=segment_sinus_beats]{segment_sinus_beats()}} on \code{egm} object
}}
}
\value{
Returns a \code{list} of \code{egm} objects. Each item is a segmentation of an \code{egm}
object, using the selected channels (if available). It will attempt to
optimize and pick the best annotations to help create consistencies between
the signal channels as possible.
}
\description{
Segmentation of electrical signal by wave specifications
}
\details{
Requires a 12-lead ECG that has been digitized, and input as an \code{egm} object.
This object must have an annotation file associated with it that contains
demarcation of \strong{P} waves, \strong{R} waves, and \strong{T} waves.
}
\section{Functions}{
\itemize{
\item \code{segment_sinus_beats()}: Identify individual sinus beats on surface ECG and
extract as individual beats, returning a list of sinus beats in the form of
the \code{egm} class. a consistent \strong{P}, \strong{R}, and \strong{T} wave amongst all
channels. If a channel does not have, for example, a visible \strong{T} wave, it
will still label it as information gained from other channels.

}}
