% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiva-package.R
\docType{package}
\name{shiva-package}
\alias{shiva}
\alias{shiva-package}
\title{shiva: Cardiac electrophysiology signal analysis}
\description{
Importing rhythm and signal data and their analysis in R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/shiva}
  \item \url{https://asshah4.github.io/shiva/}
  \item Report bugs at \url{https://github.com/asshah4/shiva/issues}
}

}
\author{
\strong{Maintainer}: Anish Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
