% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{signal}
\alias{signal}
\title{Electrophysiology signal class definition}
\usage{
signal(
  x = double(),
  label = factor(),
  source = factor(),
  color = character(),
  frequency = integer(),
  voltage = character()
)
}
\description{
Electrophysiology signal class definition
}
