% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorcardiogram.R
\name{spatial_angle}
\alias{spatial_angle}
\title{Angle between two spatial vectors}
\usage{
spatial_angle(a, b)
}
\arguments{
\item{a, b}{Numeric vectors of length three, ordered X, Y, Z.}
}
\value{
The angle in degrees, between 0 and 180.
}
\description{
Angle between two spatial vectors
}
\keyword{internal}
