% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{[.windowed}
\alias{[.windowed}
\title{Subset a windowed object}
\usage{
\method{[}{windowed}(x, i, ...)
}
\arguments{
\item{x}{A windowed object}

\item{i}{Index to subset}

\item{...}{Additional arguments passed to methods}
}
\value{
A windowed object with the specified subset of elements
}
\description{
Subset a windowed object
}
