% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows-class.R
\name{[.windows}
\alias{[.windows}
\title{Subset a \code{windows} object}
\usage{
\method{[}{windows}(x, i, ...)
}
\arguments{
\item{x}{A \code{windows} object.}

\item{i}{Index to subset.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{windows} object with the specified subset of elements.
}
\description{
Subset a \code{windows} object
}
