% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{theme_dag}
\alias{theme_dag}
\title{Personalised style for ggplot}
\usage{
theme_dag(base_size = 11, base_family = "", family = "",
  legend.position = "right")
}
\description{
\code{theme_dag} Provides customised text size and family.
}
