% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{time_normalize_windows}
\alias{time_normalize_windows}
\title{Time normalize windows to a standard length}
\usage{
time_normalize_windows(
  x,
  target_samples = 500,
  target_ms = NULL,
  interpolation_method = "linear",
  align_feature = NULL,
  channel_criteria = NULL,
  preserve_amplitude = TRUE,
  ...
)
}
\description{
Time normalize windows to a standard length
}
\keyword{internal}
