% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{validate_frequency}
\alias{validate_frequency}
\title{Validate a sampling frequency argument}
\usage{
validate_frequency(frequency, arg = "frequency")
}
\description{
Validate a sampling frequency argument
}
\keyword{internal}
