% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb.R
\name{wfdbdesc}
\alias{wfdbdesc}
\title{Describes signals based on header file contents {wfdb}}
\usage{
wfdbdesc(record, location = ".")
}
\arguments{
\item{record}{Name of the record as a string, with no extensions}

\item{location}{The directory that the target record is located within. As
this is related to the \href{https://physionet.org}{PhysioNet}, using the
location name \code{mitdb} will access the online directory for the MIT
Database.}
}
\description{
Describes signals based on header file contents {wfdb}
}
