% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window_by_rhythm}
\alias{window_by_rhythm}
\title{Window signal by rhythm patterns
Creates windows around rhythm patterns using flexible annotation criteria.}
\usage{
window_by_rhythm(
  object,
  rhythm_type = "sinus",
  onset_criteria,
  offset_criteria,
  reference_criteria = NULL,
  adjust_sample_indices = TRUE,
  ...
)
}
\description{
Window signal by rhythm patterns
Creates windows around rhythm patterns using flexible annotation criteria.
}
\keyword{internal}
