% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-annotation.R
\name{write_annotation}
\alias{write_annotation}
\title{Write an annotation into a WFDB-compatible file}
\usage{
write_annotation(data, record, annotator, record_dir = ".", wfdb_path, ...)
}
\arguments{
\item{data}{A table containing 6 columns}
}
\value{
Outputs a WFDB with the provided extension
}
\description{
Write an annotation into a WFDB-compatible file
}
