% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb.R
\name{write_wfdb}
\alias{write_wfdb}
\title{Create WFDB-compatible signal and header files from EP recording systems}
\usage{
write_wfdb(file, type, record = file, record_dir = ".", wfdb_path, ...)
}
\arguments{
\item{file}{Name of file. If it is a full path, then \code{read_location} is
ignored}

\item{type}{Type of signal data, as specified by the recording system.
Currently supports:
\itemize{
\item \emph{lspro} = Boston Scientific LabSystem Pro (Bard)
}}

\item{record}{String that will be used to name the WFDB record. Cannot
include extensions, and is not a filepath. alphanumeric characters are
acceptable, as well as hyphens (-) and underscores (_)}

\item{record_dir}{File path of directory that should be used read and write
files. Defaults to current directory.}
}
\description{
This function allows for WFDB files to be read from specific EP recording
systems, as indicated by the \strong{type} argument below.
}
