% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb.R
\name{wrsamp}
\alias{wrsamp}
\title{Create WFDB-compatible signal and header files from EP recording systems}
\usage{
wrsamp(file, type, record, read_location = ".", write_location = ".", ...)
}
\arguments{
\item{file}{File path of signal data}

\item{type}{Type of signal data, as specified by the recording system.
Currently supports:
\itemize{
\item \emph{lspro} = Boston Scientific LabSystem Pro (Bard)
}}

\item{record}{String that will be used to name the WFDB record. Cannot include extensions, and is not a filepath. alphanumeric characters are acceptable, as well as hyphens (-) and underscores (_)}
}
\description{
This function allows for WFDB files to be read from specific EP recording
systems, as indicated by the \strong{type} argument below.
}
