% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yearly_nets}
\alias{yearly_nets}
\title{Time-variant family size in Rio Negro.}
\format{A data frame with 70512 rows and 13 variables: \describe{
  \item{ego}{unique universal id}
  \item{year}{year of observation}
  \item{parents}{parents alive}
  \item{children}{children alive}
  \item{spouses}{spouses alive}
  \item{siblings}{sibligs alive}
  \item{cousins}{cousins alive}
  \item{uncles}{uncles alive}
  \item{nephews}{nephews or nieces alive}
  \item{grandparents}{grandparents alive}
  \item{grandchildren}{grandchildren alive}
  \item{relatives_alive_ext}{all relatives in extended family alive}
  \item{relatives_alive_nuc}{all relatives in nuclear family alive}
  }}
\source{
Data collected by author in Guatemala (2015-2016).
}
\usage{
yearly_nets
}
\description{
The dataset contains the number of fmaily members for Rio Negro inhabitants
over time (1891-2016).
}
\keyword{datasets}
