% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{add_annotations}
\alias{add_annotations}
\title{Add annotations to a \code{ggm} object}
\usage{
add_annotations(...)
}
\description{
The \code{add_annotations()} adds annotations to a \code{ggm} object. It is
specific to this class as it requires the output of \code{\link[=ggm]{ggm()}} to included
data stored in \code{\link[=annotation_table]{annotation_table()}}.
}
