% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{add_annotations}
\alias{add_annotations}
\title{Add annotations to a \code{ggm} object}
\usage{
add_annotations(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=ggm]{ggm}}
  \describe{
    \item{\code{data}}{Data of the \code{EGM} class, which includes header (meta) and signal
information together.}
    \item{\code{channels}}{A \code{character} vector of which channels to use. Can give
either the channel label (e.g "CS 1-2") or the recording device/catheter
type (e.g "His" or "ECG"). If no channels are selected, the default is all
channels.}
    \item{\code{time_frame}}{A time range that should be displaced given in the format
of a vector with a length of 2. The left value is the start, and right
value is the end time. This is given in seconds (decimals may be used).}
  }}
}
\description{
The \code{add_annotations()} adds annotations to a \code{ggm} object. It is
specific to this class as it requires the output of \code{\link[=ggm]{ggm()}} to included
data stored in \code{\link[=annotation_table]{annotation_table()}}.
}
