% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{add_colors}
\alias{add_colors}
\title{Add color scheme to a \code{ggm} object}
\usage{
add_colors(object, palette, mode = "light")
}
\arguments{
\item{object}{\if{html}{\out{<ggm>}} object}

\item{palette}{\if{html}{\out{<character>}} Color palette options for leads as below:
\itemize{
\item \strong{material} - material color theme
\item \strong{bw} - black or white color theme based on light/dark mode
}}

\item{mode}{\if{html}{\out{<character>}} Adjust color settings for either light or dark mode}
}
\value{
Returns an updated \verb{<ggm>} object
}
\description{
Using \code{add_colors()} is part of the theming process for a \verb{<ggm>} object,
which in turn is a visual representation of an \verb{<egm>} object. Often, the
\verb{<egm>} dataset will contain default colors based on where the signal data
was brought in from. \code{add_colors()} can allow customization of those features
to some degree based on \emph{opinionated} color palettes.
}
\details{
Currently, the color choices are individual decided based on the channel
source (e.g. lead) and are inspired by some modern palettes. The eventual
goal for this function is to accept a multitude of palette options using
heuristics similar to what is found in \code{{ggplot2}} or other graphing
packages.
}
