% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwave.R
\name{analyze_atrial_signal}
\alias{analyze_atrial_signal}
\title{Analyze F waves in atrial fibrillation ECG}
\usage{
analyze_atrial_signal(
  atrial_signal,
  frequency,
  characteristics = c("amplitude", "approximate_entropy", "dominant_frequency"),
  ...
)
}
\arguments{
\item{atrial_signal}{Numeric vector of the atrial signal}

\item{frequency}{Sampling frequency of the signal}

\item{characteristics}{Vector of characteristics to analyze}

\item{...}{Additional parameters for specific analyses}
}
\value{
A list containing the results of the requested analyses
}
\description{
Analyze F waves in atrial fibrillation ECG
}
