% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-structures.R
\name{annotation_table}
\alias{annotation_table}
\alias{is_annotation_table}
\title{Annotation Table}
\usage{
annotation_table(
  annotator = character(),
  time = numeric(),
  sample = integer(),
  type = character(),
  subtype = character(),
  channel = integer(),
  number = integer()
)

is_annotation_table(x)
}
\arguments{
\item{annotator}{String that is the name of a WFDB-compatible annotation
type, serving as the extension for the file that is written containing that
annotation. Please see \code{\link[=read_annotation]{read_annotation()}} and \code{\link[=write_annotation]{write_annotation()}} for
further details.}

\item{time}{\if{html}{\out{<double>}} A vector of}
}
\value{
A \verb{<data.table>} that has invariant columns that is compatible with
the WFDB library
}
\description{
\code{annotation_table()} modifies the \verb{<data.table>} class to work with
annotation data. The columns are of all equal length, and each row describes
a single annotation (although there may be duplicate time points).
}
\details{
The \code{annotation_table()} function creates a compatible table that can be used
with \code{\link[=write_annotation]{write_annotation()}} and \code{\link[=read_annotation]{read_annotation()}} functions.
}
