% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-data-structures.R
\name{annotation_table}
\alias{annotation_table}
\alias{is_annotation_table}
\title{Annotation Table}
\usage{
annotation_table(
  TIME = numeric(),
  SAMPLE = integer(),
  TYPE = character(),
  SUBTYPE = character(),
  CHANNEL = integer(),
  NUMBER = integer()
)

is_annotation_table(x)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{annotation_table()} modifies the \code{data.table} class to work with annotation
data.
}
