% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg.R
\name{as_ECG}
\alias{as_ECG}
\title{Convert an EGM object to an ECG object}
\usage{
as_ECG(x, ...)
}
\arguments{
\item{x}{An object of class \code{EGM}}

\item{...}{Additional arguments}
}
\value{
An object of class \code{ECG}
}
\description{
Converts a general \code{EGM} object to a specialized \code{ECG} object
for 12-lead ECG analysis.
}
