% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg.R
\name{as_ecg}
\alias{as_ecg}
\title{Convert an egm object to an ecg object}
\usage{
as_ecg(x, ...)
}
\arguments{
\item{x}{An object of class \code{egm}}

\item{...}{Additional arguments}
}
\value{
An object of class \code{ecg}
}
\description{
Converts a general \code{egm} object to a specialized \code{ecg} object
for 12-lead ECG analysis.
}
