% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwave.R
\name{calculate_dominant_frequency}
\alias{calculate_dominant_frequency}
\title{Calculate Dominant Frequency of a time series}
\usage{
calculate_dominant_frequency(x, frequency, f_min = 4, f_max = 9)
}
\arguments{
\item{x}{Numeric vector of the time series}

\item{frequency}{Sampling frequency of the signal}

\item{f_min}{Minimum frequency to consider (default 4 Hz)}

\item{f_max}{Maximum frequency to consider (default 9 Hz)}
}
\value{
Dominant Frequency in Hz
}
\description{
Calculate Dominant Frequency of a time series
}
