% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{color_channels}
\alias{color_channels}
\title{Identify the color for a channel based on palettes}
\usage{
color_channels(x, palette, mode = "light")
}
\arguments{
\item{x}{Vector of \code{character} names of requested ECG or EGM leads}

\item{palette}{A \code{character} name of the color palette options for leads as
below:
\itemize{
\item \strong{material} - material color theme
}}

\item{mode}{Adjust color settings for either light or dark mode, as
\code{character} input}
}
\value{
Vector of hex code colors as \code{character} based on the selected palette and
light/dark mode
}
\description{
This primarily restricts the colors to color-space safe options. It is
intended to be used with \code{\link[=add_colors]{add_colors()}} to provide a color scheme for the
\code{ggm} object. It has been exposed to users for custom or advanced theming
options.
}
