% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{colors}
\alias{colors}
\alias{theme_egm}
\alias{theme_egm_light}
\alias{theme_egm_dark}
\title{Theming and color options for \code{ggm} objects}
\usage{
theme_egm()

theme_egm_light()

theme_egm_dark()
}
\value{
A \code{ggm} object, with inheritance similar to
\code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The general purpose is to improve visualization of electrical signals. There
is a pattern of colors that are generally given from different recording
software, and they can be replicated to help improve visibility.
}
