% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egm-package.R
\docType{package}
\name{EGM-package}
\alias{EGM}
\alias{EGM-package}
\title{EGM: Evaluating Cardiac Electrophysiology Signals}
\description{
A system for importing electrophysiological signal, based on the 'Waveform Database (WFDB)' software package, written by Moody et al 2022 \doi{10.13026/gjvw-1m31}. A wrapper for utilizing 'WFDB' functions for reading and writing signal data, as well as functions for visualization and analysis are provided. A stable and broadly compatible class for working with signal data, supporting the reading in of cardiac electrophysiogical files such as intracardiac electrograms, is introduced.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shah-in-boots.github.io/EGM/}
  \item Report bugs at \url{https://github.com/shah-in-boots/EGM/issues}
}

}
\author{
\strong{Maintainer}: Anish S. Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID}) [copyright holder]

}
\keyword{internal}
