% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egm.R
\name{egm}
\alias{egm}
\alias{is_egm}
\title{Electrogram and electrocardiogram data class definition}
\usage{
egm(signal, header, annotation = NULL, ...)

is_egm(x)
}
\description{
Electrogram and electrocardiogram data class definition
}
