% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egm.R
\name{egm}
\alias{egm}
\alias{is_egm}
\title{Electrophysiology signal recording class definition}
\usage{
egm(header = data.table(), signal = data.table())

is_egm(x)
}
\description{
Electrophysiology signal recording class definition
}
