% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ep_models}
\alias{ep_models}
\alias{load_electrophysiology_models}
\title{Tensorflow-based electrophysiology models}
\usage{
load_electrophysiology_models(
  model_architecture,
  input_shape,
  output_distribution,
  output_type,
  version = NULL,
  ...
)
}
\arguments{
\item{model_architecture}{Neural network structure, such as \emph{CNN}. The supported types currently are:
\itemize{
\item \emph{CNN} = convolutional neural network
}}

\item{input_shape}{The dimensions of the input array for a single sample,
listed as \eqn{M \\\times N}. For example, an ECG signal of 500 samples
with 12 leads would be written as \emph{500x12}.}

\item{output_distribution}{The individual models may have a specific type of
output type. For example, if doing an ECG classification into atrial
fibrillation versus sinus (yes/no), the output type may be labeled as
\emph{AF}}

\item{version}{Some models may be have iterations that are stored. In any
models that are associated with core functions in \code{{EGM}}, several versions
will be retained. Some models may not have a version, and thus may be
\code{NULL}}

\item{...}{Additional parameters to be passed to the function}
}
\value{
A \code{SavedModel} format from \code{{keras}} and \code{{tensorflow}}
}
\description{
This function loads pre-trained electrophysiology signal
processing models trained using \code{{keras}} and \code{{tensorflow}}. This function
serves as a wrapper around \code{\link[keras:save_model_tf]{keras::load_model_tf()}}. The source for these
do not contain any personal or protected health information. The models are
generally public-domain, and can be used for research purposes.
Importantly, this function requires an internet connection to work. If
internet is not available, the function will return empty with a message.

To select a model, please use the parameters which in turn help decide what
model type is of interest. To allow for flexibility in the future, the
model names have been parameterized. For example, a model may be named
\emph{CNN_500x12_Binary_AF_v1}

The model types are detailed below.
}
