% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eps.R
\name{eps}
\alias{eps}
\alias{is_eps}
\title{Electrophysiology eps class definition}
\usage{
eps(
  x = double(),
  frequency = 1000L,
  voltage = "mV",
  label = character(),
  color = "#FFFFFF",
  ...
)

is_eps(x)
}
\description{
Electrophysiology eps class definition
}
