% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egm.R
\name{get_single_annotation}
\alias{get_single_annotation}
\title{Get a single annotation table from EGM object
If multiple annotators exist, returns the first one with a message}
\usage{
get_single_annotation(EGM_obj)
}
\description{
Get a single annotation table from EGM object
If multiple annotators exist, returns the first one with a message
}
\keyword{internal}
