% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_egm.R
\name{gg_egm}
\alias{gg_egm}
\title{Visualilzation of EGM using \code{ggplot}}
\usage{
gg_egm(
  data,
  frequency,
  channels,
  time_frame = NULL,
  annotation_channel = NULL,
  beats = NULL,
  intervals = NULL,
  hline = FALSE,
  vline = FALSE,
  ...
)
}
\arguments{
\item{data}{Long-format data for visualization. It is a combined format that includes both signal data and header information. Can be generated by the \code{\link[=melt_egm]{melt_egm()}} function.}

\item{frequency}{Integer value of the recording frequency (in Hertz).}

\item{channels}{Character vector of which channels to use. Can give either
the channel label (e.g "CS 1-2") or the recording device/catheter type (e.g
"His" or "ECG").}

\item{time_frame}{A time range that should be displaced given in the format
of a vector with a length of 2. The left value is the start, and right
value is the end time. This is given in seconds (decimals may be used).}

\item{annotation_channel}{Identifies which of the selected channels will be
used for annotation. This channel will be used for the analysis of
intervals, peaks, etc. If NULL, no annotations will be made.}

\item{beats}{A integer vector that represents the indexed beats that should
be analyzed within the plot. Each beat is indexed as \verb{1...n} from what is
within the \strong{time_frame} indicated. For example, if 5 beats are seen,
\code{beats = c(1, 5)} would select the 1st and 5th beat. This can also refer to
any large deflection that is consistent an EGM.}

\item{intervals}{A integer vector that represents the indexed intervals that
should be annotated. If NULL, no intervals will be annotated. For example,
if their are 5 beats, than there are 4 intervals between them that can be
labeled. They can be referenced by their index, e.g. \code{intervals = c(2)} to
reference the 2nd interval in a 5 beat range.}
}
\description{
Visualilzation of EGM using \code{ggplot}
}
