% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-structures.R
\name{header_table}
\alias{header_table}
\alias{is_header_table}
\title{Header Table}
\usage{
header_table(
  record_name = character(),
  number_of_channels = integer(),
  frequency = 250,
  samples = integer(),
  start_time = Sys.time(),
  ADC_saturation = integer(),
  file_name = character(),
  storage_format = 16L,
  ADC_gain = 200L,
  ADC_baseline = ADC_zero,
  ADC_units = "mV",
  ADC_resolution = 12L,
  ADC_zero = 0L,
  initial_value = ADC_zero,
  checksum = 0L,
  blocksize = 0L,
  label = character(),
  info_strings = list(),
  additional_gain = 1,
  low_pass = integer(),
  high_pass = integer(),
  color = "#000000",
  scale = integer()
)

is_header_table(x)
}
\arguments{
\item{record_name}{\if{html}{\out{<character>}} Record line information}

\item{number_of_channels}{\if{html}{\out{<integer>}} Number of signals}

\item{frequency}{\if{html}{\out{<numeric>}} Sampling frequency, 250 Hz default}

\item{samples}{\if{html}{\out{<integer>}} Number of samples}

\item{start_time}{\if{html}{\out{<POSIXct>}} Time of recording}

\item{ADC_saturation}{\if{html}{\out{<integer>}} ADC saturation}

\item{file_name}{\if{html}{\out{<character>}} Signal specific information}

\item{storage_format}{\if{html}{\out{<integer>}} Storage format, 16-bit default}

\item{ADC_gain}{\if{html}{\out{<integer>}} ADC gain, default of 200}

\item{ADC_baseline}{\if{html}{\out{<integer>}} ADC baseline, defaults to \code{ADC_zero}}

\item{ADC_units}{\if{html}{\out{<character>}} ADC units, "mV" is default}

\item{ADC_resolution}{\if{html}{\out{<integer>}} ADC resolution, default is 12}

\item{ADC_zero}{\if{html}{\out{<integer>}} ADC zero, defaults to 0}

\item{initial_value}{\if{html}{\out{<integer>}} Initial value, defaults to \code{ADC_zero} value}

\item{checksum}{\if{html}{\out{<integer>}} Checksum}

\item{blocksize}{\if{html}{\out{<integer>}} Block size}

\item{label}{\if{html}{\out{<character>}} Description}

\item{info_strings}{\if{html}{\out{<list>}} Additional information at end of file}

\item{additional_gain}{\if{html}{\out{<numeric>}} Additional gain, defaults to 1.0}

\item{low_pass}{\if{html}{\out{<integer>}} Low pass filter}

\item{high_pass}{\if{html}{\out{<integer>}} High pass filter}

\item{color}{\if{html}{\out{<character>}} Color as hexadecimal format, defaults to black}

\item{scale}{\if{html}{\out{<integer>}} Scale}
}
\description{
\code{header_table()} modifies the \code{data.table} class to work with header data.
The header data is read in from a similar format as to that of WFDB files and
should be compatible/interchangeable when writing out to disk. The details
extensively cover the type of data that is input. Generally, this function is
called by \verb{read_*_header()} functions and will generally not be called by the
end-user.
}
\section{Header file structure}{
There are three components to the header file
\enumerate{
\item \strong{Record line} that contains the following information, in the order
documented, however pieces may be missing based on different parameters. From
left to right...
\itemize{
\item Record name
\item Number of signals: represents number of segments/channels
\item Sampling frequency (optional)
\item Number of samples (optional)
\item Time: in HH:MM:SS format (optional)
\item Date: in DD/MM/YYYY (optional)
}
\item \strong{Signal specification lines} contains specifications for individual
signals, and there must be as many signal lines as there are reported by the
above record line. From left to right....
\itemize{
\item File name: usually *.dat
\item Format \if{html}{\out{<integer>}}: represents storage type, e.g. 8-bit or 16-bit
\item ADC gain: ADC units per physical unit (optional)
\itemize{
\item Baseline: corresponds to 0 physical units, sep = '*(0)" (optional)
\item Units: with '/' as a field separator e.g '*/mV' (optional)
}
\item ADC resolution \if{html}{\out{<integer>}}: bits, usually 8 or 16 (optional)
\item ADC zero: represents middle of ADC input range (optional)
\item Initial value (optional)
\item Checksum (optional)
\item Block size (optional)
\item Description: text or label information (optional)
}
\item \strong{Info strings} are unstructured lines that contains information about
the record. Usually are descriptive. Starts with initial '#' without
preceding white space at beginning of line.
}
}

