% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-data-structures.R
\name{header_table}
\alias{header_table}
\alias{is_header_table}
\title{Header Table}
\usage{
header_table(
  FILE_NAME = character(),
  NUMBER_OF_CHANNELS = integer(),
  SAMPLES = integer(),
  START_TIME = Sys.time(),
  END_TIME = Sys.time(),
  FREQUENCY = integer(),
  ADC_SATURATION = integer(),
  LABEL = character(),
  GAIN = integer(),
  ADC_GAIN = numeric(),
  LOW_PASS = integer(),
  HIGH_PASS = integer(),
  COLOR = character(),
  SCALE = integer(),
  INFO_STRINGS = list()
)

is_header_table(x)
}
\arguments{
\item{INFO}{List of characters that will be applied when writing out file}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{header_table()} modifies the \code{data.table} class to work with header data.
The header data is read in from a similar format as to that of WFDB files and
should be compatible/interchangeable when writing out to disk.
}
\section{Header file structure}{
There are three components to the header file
\enumerate{
\item \strong{Record line} that contains the following information, in the order
documented, however pieces may be missing based on different parameters. From
left to right...
\itemize{
\item Record name
\item Number of signals: represents number of segments/channels
\item Sampling frequency \link{optional}
\item Number of samples \link{optional}
\item Time: in HH:MM:SS format \link{optional}
\item Date: in DD/MM/YYYY \link{optional}
}
\item \strong{Signal specification lines} contains specifications for individual
signals, and there must be as many signal lines as there are reported by the
above record line. From left to right....
\itemize{
\item File name: (usually *.dat)
\item Format \if{html}{\out{<integer>}}: represents storage type, e.g. 8-bit or 16-bit
\item ADC gain: (ADC units per physical unit) \link{optional}
\item Baseline ADC: that corresponds to 0 physical units \link{optional}
\item Units: follows ADC gain with '/' as a field separator e.g '300/mV'
\item ADC resolution \if{html}{\out{<integer>}}: bits, usually 8 or 16 \link{optional}
\item ADC zero: represents middle of ADC input range \link{optional}
\item Initial value \link{optional}
\item Checksum \link{optional}
\item Block size \link{optional}
\item Description: text or label information \link{optional}
}
\item \strong{Info strings} are unstructured lines that contains information about
the record. Usually are descriptive. Starts with initial '#' without
preceding white space at beginning of line.
}
}

