% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-structures.R
\name{header_table}
\alias{header_table}
\alias{is_header_table}
\title{Header Table}
\usage{
header_table(
  record_name = character(),
  number_of_channels = integer(),
  frequency = 250,
  samples = integer(),
  start_time = Sys.time(),
  ADC_saturation = integer(),
  file_name = character(),
  storage_format = 16L,
  ADC_gain = 200L,
  ADC_baseline = ADC_zero,
  ADC_units = "mV",
  ADC_resolution = 12L,
  ADC_zero = 0L,
  initial_value = ADC_zero,
  checksum = 0L,
  blocksize = 0L,
  label = character(),
  info_strings = list(),
  additional_gain = 1,
  low_pass = integer(),
  high_pass = integer(),
  color = "#000000",
  scale = integer()
)

is_header_table(x)
}
\arguments{
\item{INFO}{List of characters that will be applied when writing out file}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{header_table()} modifies the \code{data.table} class to work with header data.
The header data is read in from a similar format as to that of WFDB files and
should be compatible/interchangeable when writing out to disk.
}
\section{Header file structure}{
There are three components to the header file
\enumerate{
\item \strong{Record line} that contains the following information, in the order
documented, however pieces may be missing based on different parameters. From
left to right...
\itemize{
\item Record name
\item Number of signals: represents number of segments/channels
\item Sampling frequency (optional)
\item Number of samples (optional)
\item Time: in HH:MM:SS format (optional)
\item Date: in DD/MM/YYYY (optional)
}
\item \strong{Signal specification lines} contains specifications for individual
signals, and there must be as many signal lines as there are reported by the
above record line. From left to right....
\itemize{
\item File name: usually *.dat
\item Format \if{html}{\out{<integer>}}: represents storage type, e.g. 8-bit or 16-bit
\item ADC gain: ADC units per physical unit (optional)
\itemize{
\item Baseline: corresponds to 0 physical units, sep = '*(0)" (optional)
\item Units: with '/' as a field separator e.g '*/mV' (optional)
}
\item ADC resolution \if{html}{\out{<integer>}}: bits, usually 8 or 16 (optional)
\item ADC zero: represents middle of ADC input range (optional)
\item Initial value (optional)
\item Checksum (optional)
\item Block size (optional)
\item Description: text or label information (optional)
}
\item \strong{Info strings} are unstructured lines that contains information about
the record. Usually are descriptive. Starts with initial '#' without
preceding white space at beginning of line.
}
}

