% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{interpolate_signal}
\alias{interpolate_signal}
\title{Helper function to apply interpolation}
\usage{
interpolate_signal(
  original_samples,
  original_values,
  new_samples,
  interpolation_method
)
}
\description{
Helper function to apply interpolation
}
\keyword{internal}
