% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muse.R
\name{muse}
\alias{muse}
\alias{read_muse}
\title{Read in ECG data from MUSE}
\usage{
read_muse(file)
}
\arguments{
\item{file}{An ECG file from MUSE in XML format}
}
\value{
List of 2
}
\description{
Read in ECG data from MUSE
}
