% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-annotation.R
\name{read_annotation}
\alias{read_annotation}
\title{Read WFDB-compatible annotation file}
\usage{
read_annotation(
  record,
  annotator,
  record_dir = ".",
  wfdb_path,
  begin = 0,
  end = NA_integer_,
  ...
)
}
\arguments{
\item{record}{String that will be used to name the WFDB record. Cannot
include extensions, and is not a filepath. alphanumeric characters are
acceptable, as well as hyphens (-) and underscores (_)}

\item{record_dir}{File path of directory that should be used read and write
files. Defaults to current directory.}
}
\description{
Read WFDB-compatible annotation file
}
\section{Annotation files}{
The types of annotations that are supported are described below:
\itemize{
\item atr = manually reviewed and corrected reference annotation files
\item ann = general annotator file
}
}

