% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R, R/wfdb.R
\name{read_wfdb}
\alias{read_wfdb}
\title{Read in WFDB-compatible signal files}
\usage{
read_wfdb(
  record,
  record_dir = ".",
  wfdb_path,
  begin = 0,
  end = NA_integer_,
  interval = NA_integer_,
  units = "digital",
  channels = character(),
  ...
)

read_wfdb(
  record,
  record_dir = ".",
  wfdb_path,
  begin = 0,
  end = NA_integer_,
  interval = NA_integer_,
  units = "digital",
  channels = character(),
  ...
)
}
\arguments{
\item{begin, end, interval}{Timepoint in seconds, which is converted to an
index position based on sampling frequency. The default is to start at the
beginning of the record. If \code{end} or \code{interval} are given, the earlier of
the two will be returned. The \code{end} argument gives a time index to read
until. The \code{interval} argument is the length of time past the start point.}

\item{units}{A string representing either \code{digital} (DEFAULT) or
\code{physical} units that should be returned, if available.

\if{html}{\out{<div class="sourceCode">}}\preformatted{- digital = Index in sample number, signal in integers (A/D units)

- physical = Index in elapsed time, signal in decimal voltage (e.g. mV).
This will include 1 additional row over the header/column names that
describes units
}\if{html}{\out{</div>}}}

\item{channels}{Either the signal/channel in a vector as a name or number.
Allows for duplication of signal or to re-order signal if needed.}
}
\value{
Returns a list of two objects. The first object is a matrix of
signals, with each column representing a specific channel. Each row is a
sample point. The second object is a summary of the header field, with
information on channel names/number, signal length, sampling frequency,
etc.

Data table
}
\description{
\code{read_wfdb()} reads in signal files based on the \code{rdsamp} function provided
by the original WFDB toolbox. It requires a signal file, a header file, and
an annotation file (if applicable).
}
