% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segment_ecg}
\alias{segment_ecg}
\title{Segmentation of surface ECG into individual beats}
\usage{
segment_ecg()
}
\arguments{
\item{object}{Object of the \code{egm} class, which includes header (meta), signal
information, and annotation information.}
}
\value{
Returns a single beat on all input channels (if available) as a list of \code{egm}
object. It will attempt to optimize and pick the best annotations to help
create a consistent \strong{P}, \strong{R}, and \strong{T} wave amongst all channels. If a
channel does not have, for example, a visible \strong{T} wave, it will still label
it as information gained from other channels.
}
\description{
Segmentation of surface ECG into individual beats
}
\details{
Requires a 12-lead ECG that has been digitized, and input as an \code{egm} object.
This object must have an annotation file associated with it that contains
demarcation of \strong{P} waves, \strong{R} waves, and \strong{T} waves.
}
