% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segmentation}
\alias{segmentation}
\alias{segment_sinus_beats}
\title{Segmentation of electrical signal by wave specifications}
\usage{
segmentation(object, by = "sinus", pad = 0L, center = NULL)

segment_sinus_beats(object)
}
\arguments{
\item{object}{Object of the \verb{<egm>} class, which includes header, signal
information, and annotation information.}

\item{by}{\if{html}{\out{<character>}} string naming waveform type to segment by. Options
include the following:
\itemize{
\item sinus = Will call \code{\link[=segment_sinus_beats]{segment_sinus_beats()}} on \code{egm} object
}}

\item{pad}{\if{html}{\out{<integer>}} value to offer padding of the segmented beats. The
default is \code{0L}, which means no padding will be applied. If \code{pad > 0} then
will add the baseline value (specified within the header of the signal) to
a \strong{side} of the beat. You can also choose to \strong{center} the sequence,
which will also only occur if \code{pad > 0}.}

\item{side}{\if{html}{\out{<character>}} string to specify which side of sequence to pad (or
both). Options include \code{c("left", "right", "both")}. Will only apply if
\strong{pad} is set to \verb{>0L} value. Default is \code{"right"}.}
}
\value{
Returns a list of \verb{<egm>} objects. Each item is a segmentation of an \verb{<egm>},
using the selected channels (if available). It will attempt to optimize and
pick the best annotations to help create consistencies between the signal
channels as possible.
}
\description{
Segmentation of electrical signal by wave specifications
}
\details{
Requires a 12-lead ECG that has been digitized, and input as an \code{egm} object.
This object must have an annotation file associated with it that contains
demarcation of \strong{P} waves, \strong{R} waves, and \strong{T} waves.
}
\section{Functions}{
\itemize{
\item \code{segment_sinus_beats()}: Identify individual sinus beats on surface ECG and
extract as individual beats, returning a list of sinus beats in the form of
the \code{egm} class. a consistent \strong{P}, \strong{R}, and \strong{T} wave amongst all
channels. If a channel does not have, for example, a visible \strong{T} wave, it
will still label it as information gained from other channels.

}}
