% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-data-structures.R
\name{signal_table}
\alias{signal_table}
\alias{is_signal_table}
\title{Signal Table}
\usage{
signal_table(...)

is_signal_table(x)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{signal_table()} modifies the \code{data.table} class to work with electrical
signal data. The input should be a data set of equal number of rows. It will
add a column of index positions called \code{sample} if it does not already exist.
}
