% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg.R
\name{validate_ecg_data}
\alias{validate_ecg_data}
\title{Validate that signal data represents a standard 12-lead ECG}
\usage{
validate_ecg_data(signal, header)
}
\arguments{
\item{signal}{A signal_table object}

\item{header}{A header_table object}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate that signal data represents a standard 12-lead ECG
}
\keyword{internal}
