% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-annotation.R
\name{wfdb_annotations}
\alias{wfdb_annotations}
\alias{read_annotation}
\alias{write_annotation}
\title{Read WFDB-compatible annotation file}
\usage{
read_annotation(
  record,
  record_dir = ".",
  annotator,
  wfdb_path = getOption("wfdb_path"),
  begin = 0,
  end = NULL,
  ...
)

write_annotation(
  data,
  annotator,
  record,
  record_dir = ".",
  wfdb_path = getOption("wfdb_path"),
  ...
)
}
\arguments{
\item{record}{String that will be used to name the WFDB record. Cannot
include extensions, and is not a filepath. alphanumeric characters are
acceptable, as well as hyphens (-) and underscores (_)}

\item{record_dir}{File path of directory that should be used read and write
files. Defaults to current directory.}

\item{data}{A table containing 6 columns}
}
\value{
Outputs a WFDB with the provided extension
}
\description{
The types of annotations that are supported are described below:
\itemize{
\item atr = manually reviewed and corrected reference annotation files
\item ann = general annotator file
\item ecgpuwave = \code{ecgpuwave} annotation types
}
}
\section{Annotation files}{
The following annotation file types are described below.
\subsection{ECGPUWAVE}{

\code{ecgpuwave} analyses an ECG signal from the specified record, detecting the
QRS complexes and locating the beginning, peak, and end of the P, QRS, and
ST-T waveforms. The output of ecgpuwave is written as a standard WFDB-format
annotation file associated with the specified annotator. This file can be
converted into text format using rdann or viewed using wave.

The QRS detector is based on the algorithm of Pan and Tompkins with some
improvements that make use of slope information. Optionally, QRS annotations
can be provided as input (see option -i), permitting the use of external QRS
detectors such as sqrs or manually-edited annotations (which can be created
using wave ). The waveform limit locator is based on the algorithm described
in reference 3 and evaluated in references 3 and 4.

The output annotation file contains PWAVE ("p") and TWAVE ("t")
annotations that indicate the P- and T-wave peaks, as well as QRS annotations
(NORMAL ("N") if generated by the built-in QRS detector, or copies of the
input QRS annotations if these were supplied). ecgpuwave classifies each T
wave as type 0 (normal), 1 (inverted), 2 (positive monophasic), 3 (negative
monophasic), 4 (biphasic negative-positive), or 5 (biphasic
positive-negative); this numeric classification is written into the num field
of each TWAVE annotation. The P, QRS, and T waveform onsets and ends are
marked in the output annotation file using WFON ("(") and WFOFF (")")
annotations. The num field of each WFON and WFOFF annotation designates the
type of waveform with which it is associated: 0 for a P wave, 1 for a QRS
complex, or 2 for a T wave.
}
}

