% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-io.R
\name{wfdb_io}
\alias{wfdb_io}
\alias{write_wfdb}
\alias{rewrite_wfdb}
\alias{read_wfdb}
\title{I/O of WFDB-compatible signal & header files from EP recording systems}
\usage{
write_wfdb(data, type, record = file, record_dir = ".", wfdb_path, ...)

rewrite_wfdb(file, type, record = file, record_dir = ".", wfdb_path, ...)

read_wfdb(
  record,
  record_dir = ".",
  wfdb_path,
  begin = 0,
  end = NA_integer_,
  interval = NA_integer_,
  units = c("digital", "physical"),
  channels = character(),
  ...
)
}
\arguments{
\item{data}{Name of file. If it is a full path, then \code{read_location} is
ignored}

\item{type}{Type of signal data, as specified by the recording system.
Currently supports:
\itemize{
\item \emph{lspro} = Boston Scientific LabSystem Pro (Bard)
}}

\item{record}{String that will be used to name the WFDB record. Cannot
include extensions, and is not a filepath. alphanumeric characters are
acceptable, as well as hyphens (-) and underscores (_)}

\item{record_dir}{File path of directory that should be used read and write
files. Defaults to current directory.}

\item{file}{Name of file. If it is a full path, then \code{read_location} is
ignored}

\item{begin, end, interval}{Timepoint in seconds, which is converted to an
index position based on sampling frequency. The default is to start at the
beginning of the record. If \code{end} or \code{interval} are given, the earlier of
the two will be returned. The \code{end} argument gives a time index to read
until. The \code{interval} argument is the length of time past the start point.}

\item{units}{A string representing either \code{digital} (DEFAULT) or \code{physical}
units that should be returned, if available.
\itemize{
\item digital = Index in sample number, signal in integers (A/D units)
\item physical = Index in elapsed time, signal in decimal voltage (e.g. mV).
This will include 1 additional row over the header/column names that
describes units
}}

\item{channels}{Either the signal/channel in a vector as a name or number.
Allows for duplication of signal or to re-order signal if needed.}
}
\description{
This function allows for WFDB files to be read from specific EP recording
systems, as indicated by the \strong{type} argument below.
}
