% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-io.R
\name{wfdb_io}
\alias{wfdb_io}
\alias{write_wfdb}
\alias{read_wfdb}
\alias{read_header}
\title{I/O of WFDB-compatible signal & header files from EP recording systems}
\usage{
write_wfdb(
  data,
  header = list(frequency = 250, gain = 200L, label = character()),
  info_strings = list(),
  record,
  record_dir,
  wfdb_path = getOption("wfdb_path"),
  ...
)

read_wfdb(
  record,
  record_dir = ".",
  wfdb_path = getOption("wfdb_path"),
  begin = 0,
  end = NA_integer_,
  interval = NA_integer_,
  units = "digital",
  channels = character(),
  ...
)

read_header(record, record_dir = ".", wfdb_path = getOption("wfdb_path"), ...)
}
\arguments{
\item{data}{A \code{data.frame} (or similar) that has a column that represents a
time point or index, and columns that represent signal values.}

\item{header}{A header file is a named list of parameters that will be used
to organize and describe the signal input from the \code{data} argument. If the
\code{type} is given, specific additional elements will be searched for, such as
the low or high pass filters, colors, or other signal attributes. At
minimum, the following elements are required (as cannot be calculated):
\itemize{
\item frequency = sample frequency in Hertz \if{html}{\out{<integer>}}
\item label = vector of names for each channel \if{html}{\out{<character>}}
\item start_time = date/time object \if{html}{\out{<date>}}
}}

\item{record}{String that will be used to name the WFDB record. Cannot
include extensions, and is not a filepath. alphanumeric characters are
acceptable, as well as hyphens (-) and underscores (_)}

\item{record_dir}{File path of directory that should be used read and write
files. Defaults to current directory.}

\item{begin, end, interval}{Timepoint in seconds, which is converted to an
index position based on sampling frequency. The default is to start at the
beginning of the record. If \code{end} or \code{interval} are given, the earlier of
the two will be returned. The \code{end} argument gives a time index to read
until. The \code{interval} argument is the length of time past the start point.}

\item{units}{A string representing either \code{digital} (DEFAULT) or \code{physical}
units that should be used, if available.
\itemize{
\item digital = Index in sample number, signal in integers (A/D units)
\item physical = Index in elapsed time, signal in decimal voltage (e.g. mV).
This will \strong{include 1 additional row over the header/column names} that
describes units
}}

\item{channels}{Either the signal/channel in a vector as a name or number.
Allows for duplication of signal or to re-order signal if needed.}

\item{type}{Type of signal data, as specified by the recording system.
Currently supports:
\itemize{
\item \emph{lspro} = Boston Scientific LabSystem Pro (Bard)
\item \emph{muse} = GE MUSE
}}
}
\description{
This function allows for WFDB files to be read from specific EP recording
systems, as indicated by the \strong{type} argument below.
}
