% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{windowed}
\alias{windowed}
\title{Create a windowed object containing a list of egm segments}
\usage{
windowed(x = list(), method = "rhythm", source_record = character(), ...)
}
\arguments{
\item{x}{A list of \code{egm} objects}

\item{method}{The windowing method used to create the list}

\item{source_record}{The name of the original record}

\item{...}{Additional arguments passed to methods}
}
\value{
An object of class \code{windowed} which inherits from \code{list}
}
\description{
\code{windowed} objects are lists of \code{egm} objects that represent segments or windows
of the original signal. This allows for specialized methods to be applied to
collections of signal windows.
}
